split algorithm and sampler
